refactor: protoをinterfaces層から外す
